Inline dragMagnifier helper
